requirements.txt
alembic==1.0.7
firebase-admin==2.15.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
PyJWT==1.7.1
pytest-flask==0.14.0
PyMySQL==0.9.3
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
redis==3.0.1
requests==2.21.0
SQLAlchemy==1.2.17
gunicorn==19.9.0
gevent==1.4.0
flask-cors==3.0.7
pylint==2.3.1
marshmallow==3.0.0b7